+++ include/core/parseutil.h
@@ -48,8 +48,10 @@ public:
     QStringList readCArray(const QString &text, const QString &label);
     QMap<QString, QString> readNamedIndexCArray(const QString &text, const QString &label);
     QString readCIncbin(const QString &text, const QString &label);
+    QStringList readCIncbinArray(const QString &filename, const QString &label);
     QMap<QString, int> readCDefines(const QString &filename, const QStringList &prefixes, QMap<QString, int> = { });
     QStringList readCDefinesSorted(const QString&, const QStringList&, const QMap<QString, int>& = { });
+    QMap<QString, QHash<QString, QString>> readCStructs(const QString &, const QString & = "", const QHash<int, QString> = { });
     QList<QStringList> getLabelMacros(const QList<QStringList>&, const QString&);
     QStringList getLabelValues(const QList<QStringList>&, const QString&);
     bool tryParseJsonFile(QJsonDocument *out, const QString &filepath);
@@ -69,6 +71,7 @@ public:
     static QString removeLineComments(QString text, const QStringList &commentSymbols);
 
     static QStringList splitShellCommand(QStringView command);
+    static bool gameStringToBool(QString gameString, bool * ok = nullptr);
 
 private:
     QString root;